Move the teams api requests to its own file
